Cassandra seeding complete
